Split up Io into Input, Output and File classes
The previous Io class was a design nightmare. (due to historical
reasosns) This commit is the first step of a reactoring. There are now
dedicated Input and Output classes.

In addition to this "direct" access to GET, POST and SERVER values are
possible.

The handling of remote and uploaded files was wrapped an a simple class
hierarchy.